+++ packages/edit-site/src/components/page/index.js
@@ -4,6 +4,12 @@
 import classnames from 'classnames';
 
 /**
+ * WordPress dependencies
+ */
+import { NavigableRegion } from '@wordpress/interface';
+import { EditorSnackbars } from '@wordpress/editor';
+
+/**
  * Internal dependencies
  */
 import Header from './header';
@@ -14,19 +20,23 @@ export default function Page( {
 	actions,
 	children,
 	className,
+	hideTitleFromUI = false,
 } ) {
 	const classes = classnames( 'edit-site-page', className );
 
 	return (
-		<div className={ classes }>
-			{ title && (
+		<NavigableRegion className={ classes } ariaLabel={ title }>
+			{ ! hideTitleFromUI && title && (
 				<Header
 					title={ title }
 					subTitle={ subTitle }
 					actions={ actions }
 				/>
 			) }
-			<div className="edit-site-page-content">{ children }</div>
-		</div>
+			<div className="edit-site-page-content">
+				{ children }
+				<EditorSnackbars />
+			</div>
+		</NavigableRegion>
 	);
 }